+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 2.x branch
 ## 2.12 branch
 ### 2.12.3
-* for each user, the login log is saved.
+* for each user, the login log is saved. Each user can view his logs.
 
 ### 2.12.2
 * you can download/delete logs from the log view;